backend/requirements.txt
Django==5.2.3
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
supabase>=2.7.0
postgrest>=0.16.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=3.14.0
requests>=2.31.0